Move governor 1.2 to master

+++ src/log-decoder.c
@@ -65,7 +65,7 @@ getCurrentRestrictValue (Account * ac)
   switch (ac->info.field_level_restrict)
     {
     case CPU_PARAM:
-      return (long long)ceil(ac->current.cpu * 100.0);
+      return (long long) ceil (ac->current.cpu * 100.0);
     case READ_PARAM:
       return ac->current.read;
     case WRITE_PARAM:
@@ -82,7 +82,7 @@ getShortRestrictValue (Account * ac)
   switch (ac->info.field_level_restrict)
     {
     case CPU_PARAM:
-      return (long long)ceil(ac->short_average.cpu * 100.0);
+      return (long long) ceil (ac->short_average.cpu * 100.0);
     case READ_PARAM:
       return ac->short_average.read;
     case WRITE_PARAM:
@@ -99,7 +99,7 @@ getMidRestrictValue (Account * ac)
   switch (ac->info.field_level_restrict)
     {
     case CPU_PARAM:
-      return (long long)ceil(ac->mid_average.cpu * 100.0);
+      return (long long) ceil (ac->mid_average.cpu * 100.0);
     case READ_PARAM:
       return ac->mid_average.read;
     case WRITE_PARAM:
@@ -116,7 +116,7 @@ getLongRestrictValue (Account * ac)
   switch (ac->info.field_level_restrict)
     {
     case CPU_PARAM:
-      return (long long)ceil(ac->long_average.cpu * 100.0);
+      return (long long) ceil (ac->long_average.cpu * 100.0);
     case READ_PARAM:
       return ac->long_average.read;
     case WRITE_PARAM:
@@ -143,54 +143,60 @@ getLimitValue (Account * ac, stats_limit_cfg * lm)
     }
 }
 
-void getPeriodName(char *ch, Account * ac){
-	switch (ac->info.field_restrict)
-		{
-		case CURRENT_PERIOD:
-		  strcpy (ch, "current value");
-		  break;
-		case SHORT_PERIOD:
-		  strcpy (ch, "short av.value");
-		  break;
-		case MID_PERIOD:
-		  strcpy (ch, "middle av.value");
-		  break;
-		case LONG_PERIOD:
-		  strcpy (ch, "long av.value");
-		  break;
-		default:
-		  strcpy (ch, "undef");
-		  break;
-		};
+void
+getPeriodName (char *ch, Account * ac)
+{
+  switch (ac->info.field_restrict)
+    {
+    case CURRENT_PERIOD:
+      strcpy (ch, "current value");
+      break;
+    case SHORT_PERIOD:
+      strcpy (ch, "short av.value");
+      break;
+    case MID_PERIOD:
+      strcpy (ch, "middle av.value");
+      break;
+    case LONG_PERIOD:
+      strcpy (ch, "long av.value");
+      break;
+    default:
+      strcpy (ch, "undef");
+      break;
+    };
 }
 
-void getParamName(char *buffer, Account * ac){
-	switch (ac->info.field_level_restrict)
-		{
-		case CPU_PARAM:
-		  strcpy (buffer, "cpu");
-		  break;
-		case READ_PARAM:
-		  strcpy (buffer, "read");
-		  break;
-		case WRITE_PARAM:
-		  strcpy(buffer, "write");
-		  break;
-		default:
-		  strcpy(buffer, "manual");
-		  break;
-		}
+void
+getParamName (char *buffer, Account * ac)
+{
+  switch (ac->info.field_level_restrict)
+    {
+    case CPU_PARAM:
+      strcpy (buffer, "cpu");
+      break;
+    case READ_PARAM:
+      strcpy (buffer, "read");
+      break;
+    case WRITE_PARAM:
+      strcpy (buffer, "write");
+      break;
+    default:
+      strcpy (buffer, "manual");
+      break;
+    }
 }
 
-void insertSystemInfo(char *buffer) {
-	char loadavg[GETSYSINFO_MAXFILECONTENT];
-	char vmstat[GETSYSINFO_MAXFILECONTENT];
-	char innerBuffer[_DBGOVERNOR_BUFFER_8192];
-	getloadavggov(loadavg);
-	getvmstat(vmstat);
-	snprintf(innerBuffer, _DBGOVERNOR_BUFFER_8192, "%s loadavg(%s) vmstat(%s)",
-			buffer, loadavg, vmstat);
-	strlcpy(buffer, innerBuffer, _DBGOVERNOR_BUFFER_8192);
+void
+insertSystemInfo (char *buffer)
+{
+  char loadavg[GETSYSINFO_MAXFILECONTENT];
+  char vmstat[GETSYSINFO_MAXFILECONTENT];
+  char innerBuffer[_DBGOVERNOR_BUFFER_8192];
+  getloadavggov (loadavg);
+  getvmstat (vmstat);
+  snprintf (innerBuffer, _DBGOVERNOR_BUFFER_8192, "%s loadavg(%s) vmstat(%s)",
+	    buffer, loadavg, vmstat);
+  strlcpy (buffer, innerBuffer, _DBGOVERNOR_BUFFER_8192);
 }
 
 void
@@ -200,7 +206,7 @@ prepareRestrictDescription (char *buffer, Account * ac,
   char ch[32];
   char varName[_DBGOVERNOR_BUFFER_128];
   strcpy (buffer, "");
-  if (ac->info.field_restrict==NO_PERIOD)
+  if (ac->info.field_restrict == NO_PERIOD)
     {
       strcpy (buffer, "unrestrict");
       if (cfg->restrict_format >= 2)
@@ -210,27 +216,30 @@ prepareRestrictDescription (char *buffer, Account * ac,
   else
     {
 
-	  getPeriodName(ch, ac);
-	  getParamName(varName, ac);
-	  sprintf (buffer,
-		   "%s LIMIT_ENFORCED period %s, field %s value %llu/limit %ld restrict level %d",
-		   ac->id, ch, varName, getRestrictValue (ac),
-		   getLimitValue (ac, limit), ac->restricted + 1);
-      if (cfg->restrict_format >= 2){
-    	  insertSystemInfo (buffer);
-      }
+      getPeriodName (ch, ac);
+      getParamName (varName, ac);
+      sprintf (buffer,
+	       "%s LIMIT_ENFORCED period %s, field %s value %llu/limit %ld restrict level %d",
+	       ac->id, ch, varName,
+	       (unsigned long long) getRestrictValue (ac), getLimitValue (ac,
+									  limit),
+	       ac->restricted + 1);
+      if (cfg->restrict_format >= 2)
+	{
+	  insertSystemInfo (buffer);
+	}
       return;
     }
 }
 
 void
 prepareRestrictDescriptionLimit (char *buffer, Account * ac,
-			    stats_limit_cfg * limit)
+				 stats_limit_cfg * limit)
 {
   char ch[32];
   char varName[_DBGOVERNOR_BUFFER_128];
   strcpy (buffer, "");
-  if (ac->info.field_restrict==NO_PERIOD)
+  if (ac->info.field_restrict == NO_PERIOD)
     {
       strcpy (buffer, "unrestrict");
       if (cfg->restrict_format >= 2)
@@ -240,33 +249,35 @@ prepareRestrictDescriptionLimit (char *buffer, Account * ac,
   else
     {
 
-	  getPeriodName(ch, ac);
-	  getParamName(varName, ac);
-	  sprintf (buffer,
-		   "%s LIMIT_ENFORCED period %s, field %s value %llu/limit %ld",
-		   ac->id, ch, varName, getRestrictValue (ac),
-		   getLimitValue (ac, limit));
-      if (cfg->restrict_format >= 2){
-    	  insertSystemInfo (buffer);
-      }
+      getPeriodName (ch, ac);
+      getParamName (varName, ac);
+      sprintf (buffer,
+	       "%s LIMIT_ENFORCED period %s, field %s value %llu/limit %ld",
+	       ac->id, ch, varName,
+	       (unsigned long long) getRestrictValue (ac), getLimitValue (ac,
+									  limit));
+      if (cfg->restrict_format >= 2)
+	{
+	  insertSystemInfo (buffer);
+	}
       return;
     }
 }
 
 /*Получить соответсвующий периоду список параметров, т.н дамп*/
 stats_limit *
-getRestrictDump(Account * ac)
+getRestrictDump (Account * ac)
 {
   switch (ac->info.field_restrict)
     {
     case CURRENT_PERIOD:
-      return (stats_limit *)&ac->current;
+      return (stats_limit *) & ac->current;
     case SHORT_PERIOD:
-      return (stats_limit *)&ac->short_average;
+      return (stats_limit *) & ac->short_average;
     case MID_PERIOD:
-      return (stats_limit *)&ac->mid_average;
+      return (stats_limit *) & ac->mid_average;
     case LONG_PERIOD:
-      return (stats_limit *)&ac->long_average;
+      return (stats_limit *) & ac->long_average;
     };
-  return (stats_limit *)NULL;
+  return (stats_limit *) NULL;
 }
